fix emoji
